Reorder project files and start moving functions out
