--- CrusherChess/CrusherChess/CrusherChess.vcxproj.user
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
-    <LocalDebuggerCommand>$(TargetPath)</LocalDebuggerCommand>
-    <WebServiceDebuggerHttpUrl>localhost:8080</WebServiceDebuggerHttpUrl>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
-    <LocalDebuggerCommand>$(TargetPath)</LocalDebuggerCommand>
-    <DebuggerFlavor>WebServiceDebugger</DebuggerFlavor>
-    <WebServiceDebuggerHttpUrl>localhost:8080</WebServiceDebuggerHttpUrl>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
-    <LocalDebuggerCommand>$(TargetPath)</LocalDebuggerCommand>
-    <DebuggerFlavor>WebServiceDebugger</DebuggerFlavor>
-    <WebServiceDebuggerHttpUrl>localhost:8080</WebServiceDebuggerHttpUrl>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <LocalDebuggerCommand>$(TargetPath)</LocalDebuggerCommand>
-    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
-    <WebServiceDebuggerHttpUrl>localhost:8080</WebServiceDebuggerHttpUrl>
-  </PropertyGroup>
-</Project>
