+++ CrusherChess/src/Debug.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void print_bitboard(unsigned long long int bitboard);
+void print_magic_numbers();
+
