+++ CrusherChess/src/macros.h
@@ -1,0 +1,1 @@
+#pragma once
